页面提交和Dao

+++ server/src/main/java/com/example/server/mapper/po/BasicInfoPo.java
@@ -23,7 +23,7 @@ public class BasicInfoPo {
     private String name;
 
     @Column(name = "yield")
-    private Integer yield;
+    private Double yield;
 
     @Column(name = "area")
     private Double area;
